Combined the old structure inside PackagStructure. Also deleted unneccessary methods.
